refactor(*): move util under public/lib
